+++ src/rv32_v_constopt.c
@@ -1,0 +1,599 @@
+/*
+ * rv32emu is freely redistributable under the MIT License. See the file
+ * "LICENSE" for information on usage and redistribution of this file.
+ */
+
+/* RISC-V "V" Vector Extension - constant-optimization stubs.
+ *
+ * Included from rv32_constopt.c when RV32_HAS(EXT_V) is set. Vector ops
+ * write to vector registers (rv->V[]), not GPRs (rv->X[]), so they do
+ * not invalidate constant tracking and the CONSTOPT bodies are empty.
+ * Future ops that write back to ir->rd (vmv.x.s, vcpop.m, vfirst.m,
+ * vfmv.f.s) will need info->is_constant[ir->rd] = false here.
+ */
+
+CONSTOPT(vsetvli, {})
+CONSTOPT(vsetivli, {})
+CONSTOPT(vsetvl, {})
+CONSTOPT(vle8_v, {})
+CONSTOPT(vle16_v, {})
+CONSTOPT(vle32_v, {})
+CONSTOPT(vle64_v, {})
+CONSTOPT(vlseg2e8_v, {})
+CONSTOPT(vlseg3e8_v, {})
+CONSTOPT(vlseg4e8_v, {})
+CONSTOPT(vlseg5e8_v, {})
+CONSTOPT(vlseg6e8_v, {})
+CONSTOPT(vlseg7e8_v, {})
+CONSTOPT(vlseg8e8_v, {})
+CONSTOPT(vlseg2e16_v, {})
+CONSTOPT(vlseg3e16_v, {})
+CONSTOPT(vlseg4e16_v, {})
+CONSTOPT(vlseg5e16_v, {})
+CONSTOPT(vlseg6e16_v, {})
+CONSTOPT(vlseg7e16_v, {})
+CONSTOPT(vlseg8e16_v, {})
+CONSTOPT(vlseg2e32_v, {})
+CONSTOPT(vlseg3e32_v, {})
+CONSTOPT(vlseg4e32_v, {})
+CONSTOPT(vlseg5e32_v, {})
+CONSTOPT(vlseg6e32_v, {})
+CONSTOPT(vlseg7e32_v, {})
+CONSTOPT(vlseg8e32_v, {})
+CONSTOPT(vlseg2e64_v, {})
+CONSTOPT(vlseg3e64_v, {})
+CONSTOPT(vlseg4e64_v, {})
+CONSTOPT(vlseg5e64_v, {})
+CONSTOPT(vlseg6e64_v, {})
+CONSTOPT(vlseg7e64_v, {})
+CONSTOPT(vlseg8e64_v, {})
+CONSTOPT(vl1re8_v, {})
+CONSTOPT(vl1re16_v, {})
+CONSTOPT(vl1re32_v, {})
+CONSTOPT(vl1re64_v, {})
+CONSTOPT(vl2re8_v, {})
+CONSTOPT(vl2re16_v, {})
+CONSTOPT(vl2re32_v, {})
+CONSTOPT(vl2re64_v, {})
+CONSTOPT(vl4re8_v, {})
+CONSTOPT(vl4re16_v, {})
+CONSTOPT(vl4re32_v, {})
+CONSTOPT(vl4re64_v, {})
+CONSTOPT(vl8re8_v, {})
+CONSTOPT(vl8re16_v, {})
+CONSTOPT(vl8re32_v, {})
+CONSTOPT(vl8re64_v, {})
+CONSTOPT(vlm_v, {})
+CONSTOPT(vle8ff_v, {})
+CONSTOPT(vle16ff_v, {})
+CONSTOPT(vle32ff_v, {})
+CONSTOPT(vle64ff_v, {})
+CONSTOPT(vlseg2e8ff_v, {})
+CONSTOPT(vlseg3e8ff_v, {})
+CONSTOPT(vlseg4e8ff_v, {})
+CONSTOPT(vlseg5e8ff_v, {})
+CONSTOPT(vlseg6e8ff_v, {})
+CONSTOPT(vlseg7e8ff_v, {})
+CONSTOPT(vlseg8e8ff_v, {})
+CONSTOPT(vlseg2e16ff_v, {})
+CONSTOPT(vlseg3e16ff_v, {})
+CONSTOPT(vlseg4e16ff_v, {})
+CONSTOPT(vlseg5e16ff_v, {})
+CONSTOPT(vlseg6e16ff_v, {})
+CONSTOPT(vlseg7e16ff_v, {})
+CONSTOPT(vlseg8e16ff_v, {})
+CONSTOPT(vlseg2e32ff_v, {})
+CONSTOPT(vlseg3e32ff_v, {})
+CONSTOPT(vlseg4e32ff_v, {})
+CONSTOPT(vlseg5e32ff_v, {})
+CONSTOPT(vlseg6e32ff_v, {})
+CONSTOPT(vlseg7e32ff_v, {})
+CONSTOPT(vlseg8e32ff_v, {})
+CONSTOPT(vlseg2e64ff_v, {})
+CONSTOPT(vlseg3e64ff_v, {})
+CONSTOPT(vlseg4e64ff_v, {})
+CONSTOPT(vlseg5e64ff_v, {})
+CONSTOPT(vlseg6e64ff_v, {})
+CONSTOPT(vlseg7e64ff_v, {})
+CONSTOPT(vlseg8e64ff_v, {})
+CONSTOPT(vluxei8_v, {})
+CONSTOPT(vluxei16_v, {})
+CONSTOPT(vluxei32_v, {})
+CONSTOPT(vluxei64_v, {})
+CONSTOPT(vluxseg2ei8_v, {})
+CONSTOPT(vluxseg3ei8_v, {})
+CONSTOPT(vluxseg4ei8_v, {})
+CONSTOPT(vluxseg5ei8_v, {})
+CONSTOPT(vluxseg6ei8_v, {})
+CONSTOPT(vluxseg7ei8_v, {})
+CONSTOPT(vluxseg8ei8_v, {})
+CONSTOPT(vluxseg2ei16_v, {})
+CONSTOPT(vluxseg3ei16_v, {})
+CONSTOPT(vluxseg4ei16_v, {})
+CONSTOPT(vluxseg5ei16_v, {})
+CONSTOPT(vluxseg6ei16_v, {})
+CONSTOPT(vluxseg7ei16_v, {})
+CONSTOPT(vluxseg8ei16_v, {})
+CONSTOPT(vluxseg2ei32_v, {})
+CONSTOPT(vluxseg3ei32_v, {})
+CONSTOPT(vluxseg4ei32_v, {})
+CONSTOPT(vluxseg5ei32_v, {})
+CONSTOPT(vluxseg6ei32_v, {})
+CONSTOPT(vluxseg7ei32_v, {})
+CONSTOPT(vluxseg8ei32_v, {})
+CONSTOPT(vluxseg2ei64_v, {})
+CONSTOPT(vluxseg3ei64_v, {})
+CONSTOPT(vluxseg4ei64_v, {})
+CONSTOPT(vluxseg5ei64_v, {})
+CONSTOPT(vluxseg6ei64_v, {})
+CONSTOPT(vluxseg7ei64_v, {})
+CONSTOPT(vluxseg8ei64_v, {})
+CONSTOPT(vlse8_v, {})
+CONSTOPT(vlse16_v, {})
+CONSTOPT(vlse32_v, {})
+CONSTOPT(vlse64_v, {})
+CONSTOPT(vlsseg2e8_v, {})
+CONSTOPT(vlsseg3e8_v, {})
+CONSTOPT(vlsseg4e8_v, {})
+CONSTOPT(vlsseg5e8_v, {})
+CONSTOPT(vlsseg6e8_v, {})
+CONSTOPT(vlsseg7e8_v, {})
+CONSTOPT(vlsseg8e8_v, {})
+CONSTOPT(vlsseg2e16_v, {})
+CONSTOPT(vlsseg3e16_v, {})
+CONSTOPT(vlsseg4e16_v, {})
+CONSTOPT(vlsseg5e16_v, {})
+CONSTOPT(vlsseg6e16_v, {})
+CONSTOPT(vlsseg7e16_v, {})
+CONSTOPT(vlsseg8e16_v, {})
+CONSTOPT(vlsseg2e32_v, {})
+CONSTOPT(vlsseg3e32_v, {})
+CONSTOPT(vlsseg4e32_v, {})
+CONSTOPT(vlsseg5e32_v, {})
+CONSTOPT(vlsseg6e32_v, {})
+CONSTOPT(vlsseg7e32_v, {})
+CONSTOPT(vlsseg8e32_v, {})
+CONSTOPT(vlsseg2e64_v, {})
+CONSTOPT(vlsseg3e64_v, {})
+CONSTOPT(vlsseg4e64_v, {})
+CONSTOPT(vlsseg5e64_v, {})
+CONSTOPT(vlsseg6e64_v, {})
+CONSTOPT(vlsseg7e64_v, {})
+CONSTOPT(vlsseg8e64_v, {})
+CONSTOPT(vloxei8_v, {})
+CONSTOPT(vloxei16_v, {})
+CONSTOPT(vloxei32_v, {})
+CONSTOPT(vloxei64_v, {})
+CONSTOPT(vloxseg2ei8_v, {})
+CONSTOPT(vloxseg3ei8_v, {})
+CONSTOPT(vloxseg4ei8_v, {})
+CONSTOPT(vloxseg5ei8_v, {})
+CONSTOPT(vloxseg6ei8_v, {})
+CONSTOPT(vloxseg7ei8_v, {})
+CONSTOPT(vloxseg8ei8_v, {})
+CONSTOPT(vloxseg2ei16_v, {})
+CONSTOPT(vloxseg3ei16_v, {})
+CONSTOPT(vloxseg4ei16_v, {})
+CONSTOPT(vloxseg5ei16_v, {})
+CONSTOPT(vloxseg6ei16_v, {})
+CONSTOPT(vloxseg7ei16_v, {})
+CONSTOPT(vloxseg8ei16_v, {})
+CONSTOPT(vloxseg2ei32_v, {})
+CONSTOPT(vloxseg3ei32_v, {})
+CONSTOPT(vloxseg4ei32_v, {})
+CONSTOPT(vloxseg5ei32_v, {})
+CONSTOPT(vloxseg6ei32_v, {})
+CONSTOPT(vloxseg7ei32_v, {})
+CONSTOPT(vloxseg8ei32_v, {})
+CONSTOPT(vloxseg2ei64_v, {})
+CONSTOPT(vloxseg3ei64_v, {})
+CONSTOPT(vloxseg4ei64_v, {})
+CONSTOPT(vloxseg5ei64_v, {})
+CONSTOPT(vloxseg6ei64_v, {})
+CONSTOPT(vloxseg7ei64_v, {})
+CONSTOPT(vloxseg8ei64_v, {})
+CONSTOPT(vse8_v, {})
+CONSTOPT(vse16_v, {})
+CONSTOPT(vse32_v, {})
+CONSTOPT(vse64_v, {})
+CONSTOPT(vsseg2e8_v, {})
+CONSTOPT(vsseg3e8_v, {})
+CONSTOPT(vsseg4e8_v, {})
+CONSTOPT(vsseg5e8_v, {})
+CONSTOPT(vsseg6e8_v, {})
+CONSTOPT(vsseg7e8_v, {})
+CONSTOPT(vsseg8e8_v, {})
+CONSTOPT(vsseg2e16_v, {})
+CONSTOPT(vsseg3e16_v, {})
+CONSTOPT(vsseg4e16_v, {})
+CONSTOPT(vsseg5e16_v, {})
+CONSTOPT(vsseg6e16_v, {})
+CONSTOPT(vsseg7e16_v, {})
+CONSTOPT(vsseg8e16_v, {})
+CONSTOPT(vsseg2e32_v, {})
+CONSTOPT(vsseg3e32_v, {})
+CONSTOPT(vsseg4e32_v, {})
+CONSTOPT(vsseg5e32_v, {})
+CONSTOPT(vsseg6e32_v, {})
+CONSTOPT(vsseg7e32_v, {})
+CONSTOPT(vsseg8e32_v, {})
+CONSTOPT(vsseg2e64_v, {})
+CONSTOPT(vsseg3e64_v, {})
+CONSTOPT(vsseg4e64_v, {})
+CONSTOPT(vsseg5e64_v, {})
+CONSTOPT(vsseg6e64_v, {})
+CONSTOPT(vsseg7e64_v, {})
+CONSTOPT(vsseg8e64_v, {})
+CONSTOPT(vs1r_v, {})
+CONSTOPT(vs2r_v, {})
+CONSTOPT(vs4r_v, {})
+CONSTOPT(vs8r_v, {})
+CONSTOPT(vsm_v, {})
+CONSTOPT(vsuxei8_v, {})
+CONSTOPT(vsuxei16_v, {})
+CONSTOPT(vsuxei32_v, {})
+CONSTOPT(vsuxei64_v, {})
+CONSTOPT(vsuxseg2ei8_v, {})
+CONSTOPT(vsuxseg3ei8_v, {})
+CONSTOPT(vsuxseg4ei8_v, {})
+CONSTOPT(vsuxseg5ei8_v, {})
+CONSTOPT(vsuxseg6ei8_v, {})
+CONSTOPT(vsuxseg7ei8_v, {})
+CONSTOPT(vsuxseg8ei8_v, {})
+CONSTOPT(vsuxseg2ei16_v, {})
+CONSTOPT(vsuxseg3ei16_v, {})
+CONSTOPT(vsuxseg4ei16_v, {})
+CONSTOPT(vsuxseg5ei16_v, {})
+CONSTOPT(vsuxseg6ei16_v, {})
+CONSTOPT(vsuxseg7ei16_v, {})
+CONSTOPT(vsuxseg8ei16_v, {})
+CONSTOPT(vsuxseg2ei32_v, {})
+CONSTOPT(vsuxseg3ei32_v, {})
+CONSTOPT(vsuxseg4ei32_v, {})
+CONSTOPT(vsuxseg5ei32_v, {})
+CONSTOPT(vsuxseg6ei32_v, {})
+CONSTOPT(vsuxseg7ei32_v, {})
+CONSTOPT(vsuxseg8ei32_v, {})
+CONSTOPT(vsuxseg2ei64_v, {})
+CONSTOPT(vsuxseg3ei64_v, {})
+CONSTOPT(vsuxseg4ei64_v, {})
+CONSTOPT(vsuxseg5ei64_v, {})
+CONSTOPT(vsuxseg6ei64_v, {})
+CONSTOPT(vsuxseg7ei64_v, {})
+CONSTOPT(vsuxseg8ei64_v, {})
+CONSTOPT(vsse8_v, {})
+CONSTOPT(vsse16_v, {})
+CONSTOPT(vsse32_v, {})
+CONSTOPT(vsse64_v, {})
+CONSTOPT(vssseg2e8_v, {})
+CONSTOPT(vssseg3e8_v, {})
+CONSTOPT(vssseg4e8_v, {})
+CONSTOPT(vssseg5e8_v, {})
+CONSTOPT(vssseg6e8_v, {})
+CONSTOPT(vssseg7e8_v, {})
+CONSTOPT(vssseg8e8_v, {})
+CONSTOPT(vssseg2e16_v, {})
+CONSTOPT(vssseg3e16_v, {})
+CONSTOPT(vssseg4e16_v, {})
+CONSTOPT(vssseg5e16_v, {})
+CONSTOPT(vssseg6e16_v, {})
+CONSTOPT(vssseg7e16_v, {})
+CONSTOPT(vssseg8e16_v, {})
+CONSTOPT(vssseg2e32_v, {})
+CONSTOPT(vssseg3e32_v, {})
+CONSTOPT(vssseg4e32_v, {})
+CONSTOPT(vssseg5e32_v, {})
+CONSTOPT(vssseg6e32_v, {})
+CONSTOPT(vssseg7e32_v, {})
+CONSTOPT(vssseg8e32_v, {})
+CONSTOPT(vssseg2e64_v, {})
+CONSTOPT(vssseg3e64_v, {})
+CONSTOPT(vssseg4e64_v, {})
+CONSTOPT(vssseg5e64_v, {})
+CONSTOPT(vssseg6e64_v, {})
+CONSTOPT(vssseg7e64_v, {})
+CONSTOPT(vssseg8e64_v, {})
+CONSTOPT(vsoxei8_v, {})
+CONSTOPT(vsoxei16_v, {})
+CONSTOPT(vsoxei32_v, {})
+CONSTOPT(vsoxei64_v, {})
+CONSTOPT(vsoxseg2ei8_v, {})
+CONSTOPT(vsoxseg3ei8_v, {})
+CONSTOPT(vsoxseg4ei8_v, {})
+CONSTOPT(vsoxseg5ei8_v, {})
+CONSTOPT(vsoxseg6ei8_v, {})
+CONSTOPT(vsoxseg7ei8_v, {})
+CONSTOPT(vsoxseg8ei8_v, {})
+CONSTOPT(vsoxseg2ei16_v, {})
+CONSTOPT(vsoxseg3ei16_v, {})
+CONSTOPT(vsoxseg4ei16_v, {})
+CONSTOPT(vsoxseg5ei16_v, {})
+CONSTOPT(vsoxseg6ei16_v, {})
+CONSTOPT(vsoxseg7ei16_v, {})
+CONSTOPT(vsoxseg8ei16_v, {})
+CONSTOPT(vsoxseg2ei32_v, {})
+CONSTOPT(vsoxseg3ei32_v, {})
+CONSTOPT(vsoxseg4ei32_v, {})
+CONSTOPT(vsoxseg5ei32_v, {})
+CONSTOPT(vsoxseg6ei32_v, {})
+CONSTOPT(vsoxseg7ei32_v, {})
+CONSTOPT(vsoxseg8ei32_v, {})
+CONSTOPT(vsoxseg2ei64_v, {})
+CONSTOPT(vsoxseg3ei64_v, {})
+CONSTOPT(vsoxseg4ei64_v, {})
+CONSTOPT(vsoxseg5ei64_v, {})
+CONSTOPT(vsoxseg6ei64_v, {})
+CONSTOPT(vsoxseg7ei64_v, {})
+CONSTOPT(vsoxseg8ei64_v, {})
+CONSTOPT(vadd_vv, {})
+CONSTOPT(vadd_vx, {})
+CONSTOPT(vadd_vi, {})
+CONSTOPT(vsub_vv, {})
+CONSTOPT(vsub_vx, {})
+CONSTOPT(vrsub_vx, {})
+CONSTOPT(vrsub_vi, {})
+CONSTOPT(vminu_vv, {})
+CONSTOPT(vminu_vx, {})
+CONSTOPT(vmin_vv, {})
+CONSTOPT(vmin_vx, {})
+CONSTOPT(vmaxu_vv, {})
+CONSTOPT(vmaxu_vx, {})
+CONSTOPT(vmax_vv, {})
+CONSTOPT(vmax_vx, {})
+CONSTOPT(vand_vv, {})
+CONSTOPT(vand_vx, {})
+CONSTOPT(vand_vi, {})
+CONSTOPT(vor_vv, {})
+CONSTOPT(vor_vx, {})
+CONSTOPT(vor_vi, {})
+CONSTOPT(vxor_vv, {})
+CONSTOPT(vxor_vx, {})
+CONSTOPT(vxor_vi, {})
+CONSTOPT(vrgather_vv, {})
+CONSTOPT(vrgather_vx, {})
+CONSTOPT(vrgather_vi, {})
+CONSTOPT(vslideup_vx, {})
+CONSTOPT(vslideup_vi, {})
+CONSTOPT(vrgatherei16_vv, {})
+CONSTOPT(vslidedown_vx, {})
+CONSTOPT(vslidedown_vi, {})
+CONSTOPT(vadc_vvm, {})
+CONSTOPT(vadc_vxm, {})
+CONSTOPT(vadc_vim, {})
+CONSTOPT(vmadc_vv, {})
+CONSTOPT(vmadc_vx, {})
+CONSTOPT(vmadc_vi, {})
+CONSTOPT(vsbc_vvm, {})
+CONSTOPT(vsbc_vxm, {})
+CONSTOPT(vmsbc_vv, {})
+CONSTOPT(vmsbc_vx, {})
+CONSTOPT(vmerge_vvm, {})
+CONSTOPT(vmerge_vxm, {})
+CONSTOPT(vmerge_vim, {})
+CONSTOPT(vmv_v_v, {})
+CONSTOPT(vmv_v_x, {})
+CONSTOPT(vmv_v_i, {})
+CONSTOPT(vmseq_vv, {})
+CONSTOPT(vmseq_vx, {})
+CONSTOPT(vmseq_vi, {})
+CONSTOPT(vmsne_vv, {})
+CONSTOPT(vmsne_vx, {})
+CONSTOPT(vmsne_vi, {})
+CONSTOPT(vmsltu_vv, {})
+CONSTOPT(vmsltu_vx, {})
+CONSTOPT(vmslt_vv, {})
+CONSTOPT(vmslt_vx, {})
+CONSTOPT(vmsleu_vv, {})
+CONSTOPT(vmsleu_vx, {})
+CONSTOPT(vmsleu_vi, {})
+CONSTOPT(vmsle_vv, {})
+CONSTOPT(vmsle_vx, {})
+CONSTOPT(vmsle_vi, {})
+CONSTOPT(vmsgtu_vx, {})
+CONSTOPT(vmsgtu_vi, {})
+CONSTOPT(vmsgt_vx, {})
+CONSTOPT(vmsgt_vi, {})
+CONSTOPT(vsaddu_vv, {})
+CONSTOPT(vsaddu_vx, {})
+CONSTOPT(vsaddu_vi, {})
+CONSTOPT(vsadd_vv, {})
+CONSTOPT(vsadd_vx, {})
+CONSTOPT(vsadd_vi, {})
+CONSTOPT(vssubu_vv, {})
+CONSTOPT(vssubu_vx, {})
+CONSTOPT(vssub_vv, {})
+CONSTOPT(vssub_vx, {})
+CONSTOPT(vsll_vv, {})
+CONSTOPT(vsll_vx, {})
+CONSTOPT(vsll_vi, {})
+CONSTOPT(vsmul_vv, {})
+CONSTOPT(vsmul_vx, {})
+CONSTOPT(vsrl_vv, {})
+CONSTOPT(vsrl_vx, {})
+CONSTOPT(vsrl_vi, {})
+CONSTOPT(vsra_vv, {})
+CONSTOPT(vsra_vx, {})
+CONSTOPT(vsra_vi, {})
+CONSTOPT(vssrl_vv, {})
+CONSTOPT(vssrl_vx, {})
+CONSTOPT(vssrl_vi, {})
+CONSTOPT(vssra_vv, {})
+CONSTOPT(vssra_vx, {})
+CONSTOPT(vssra_vi, {})
+CONSTOPT(vnsrl_wv, {})
+CONSTOPT(vnsrl_wx, {})
+CONSTOPT(vnsrl_wi, {})
+CONSTOPT(vnsra_wv, {})
+CONSTOPT(vnsra_wx, {})
+CONSTOPT(vnsra_wi, {})
+CONSTOPT(vnclipu_wv, {})
+CONSTOPT(vnclipu_wx, {})
+CONSTOPT(vnclipu_wi, {})
+CONSTOPT(vnclip_wv, {})
+CONSTOPT(vnclip_wx, {})
+CONSTOPT(vnclip_wi, {})
+CONSTOPT(vwredsumu_vs, {})
+CONSTOPT(vwredsum_vs, {})
+CONSTOPT(vredsum_vs, {})
+CONSTOPT(vredand_vs, {})
+CONSTOPT(vredor_vs, {})
+CONSTOPT(vredxor_vs, {})
+CONSTOPT(vredminu_vs, {})
+CONSTOPT(vredmin_vs, {})
+CONSTOPT(vredmaxu_vs, {})
+CONSTOPT(vredmax_vs, {})
+CONSTOPT(vaaddu_vv, {})
+CONSTOPT(vaaddu_vx, {})
+CONSTOPT(vaadd_vv, {})
+CONSTOPT(vaadd_vx, {})
+CONSTOPT(vasubu_vv, {})
+CONSTOPT(vasubu_vx, {})
+CONSTOPT(vasub_vv, {})
+CONSTOPT(vasub_vx, {})
+CONSTOPT(vslide1up_vx, {})
+CONSTOPT(vslide1down_vx, {})
+CONSTOPT(vcompress_vm, {})
+CONSTOPT(vmandn_mm, {})
+CONSTOPT(vmand_mm, {})
+CONSTOPT(vmor_mm, {})
+CONSTOPT(vmxor_mm, {})
+CONSTOPT(vmorn_mm, {})
+CONSTOPT(vmnand_mm, {})
+CONSTOPT(vmnor_mm, {})
+CONSTOPT(vmxnor_mm, {})
+CONSTOPT(vdivu_vv, {})
+CONSTOPT(vdivu_vx, {})
+CONSTOPT(vdiv_vv, {})
+CONSTOPT(vdiv_vx, {})
+CONSTOPT(vremu_vv, {})
+CONSTOPT(vremu_vx, {})
+CONSTOPT(vrem_vv, {})
+CONSTOPT(vrem_vx, {})
+CONSTOPT(vmulhu_vv, {})
+CONSTOPT(vmulhu_vx, {})
+CONSTOPT(vmul_vv, {})
+CONSTOPT(vmul_vx, {})
+CONSTOPT(vmulhsu_vv, {})
+CONSTOPT(vmulhsu_vx, {})
+CONSTOPT(vmulh_vv, {})
+CONSTOPT(vmulh_vx, {})
+CONSTOPT(vmadd_vv, {})
+CONSTOPT(vmadd_vx, {})
+CONSTOPT(vnmsub_vv, {})
+CONSTOPT(vnmsub_vx, {})
+CONSTOPT(vmacc_vv, {})
+CONSTOPT(vmacc_vx, {})
+CONSTOPT(vnmsac_vv, {})
+CONSTOPT(vnmsac_vx, {})
+CONSTOPT(vwaddu_vv, {})
+CONSTOPT(vwaddu_vx, {})
+CONSTOPT(vwadd_vv, {})
+CONSTOPT(vwadd_vx, {})
+CONSTOPT(vwsubu_vv, {})
+CONSTOPT(vwsubu_vx, {})
+CONSTOPT(vwsub_vv, {})
+CONSTOPT(vwsub_vx, {})
+CONSTOPT(vwaddu_wv, {})
+CONSTOPT(vwaddu_wx, {})
+CONSTOPT(vwadd_wv, {})
+CONSTOPT(vwadd_wx, {})
+CONSTOPT(vwsubu_wv, {})
+CONSTOPT(vwsubu_wx, {})
+CONSTOPT(vwsub_wv, {})
+CONSTOPT(vwsub_wx, {})
+CONSTOPT(vwmulu_vv, {})
+CONSTOPT(vwmulu_vx, {})
+CONSTOPT(vwmulsu_vv, {})
+CONSTOPT(vwmulsu_vx, {})
+CONSTOPT(vwmul_vv, {})
+CONSTOPT(vwmul_vx, {})
+CONSTOPT(vwmaccu_vv, {})
+CONSTOPT(vwmaccu_vx, {})
+CONSTOPT(vwmacc_vv, {})
+CONSTOPT(vwmacc_vx, {})
+CONSTOPT(vwmaccus_vx, {})
+CONSTOPT(vwmaccsu_vv, {})
+CONSTOPT(vwmaccsu_vx, {})
+CONSTOPT(vmv_s_x, {})
+CONSTOPT(vmv_x_s, {})
+CONSTOPT(vcpop_m, {})
+CONSTOPT(vfirst_m, {})
+CONSTOPT(vmsbf_m, {})
+CONSTOPT(vmsof_m, {})
+CONSTOPT(vmsif_m, {})
+CONSTOPT(viota_m, {})
+CONSTOPT(vid_v, {})
+CONSTOPT(vfadd_vv, {})
+CONSTOPT(vfadd_vf, {})
+CONSTOPT(vfredusum_vs, {})
+CONSTOPT(vfsub_vv, {})
+CONSTOPT(vfsub_vf, {})
+CONSTOPT(vfredosum_vs, {})
+CONSTOPT(vfmin_vv, {})
+CONSTOPT(vfmin_vf, {})
+CONSTOPT(vfredmin_vs, {})
+CONSTOPT(vfmax_vv, {})
+CONSTOPT(vfmax_vf, {})
+CONSTOPT(vfredmax_vs, {})
+CONSTOPT(vfsgnj_vv, {})
+CONSTOPT(vfsgnj_vf, {})
+CONSTOPT(vfsgnjn_vv, {})
+CONSTOPT(vfsgnjn_vf, {})
+CONSTOPT(vfsgnjx_vv, {})
+CONSTOPT(vfsgnjx_vf, {})
+CONSTOPT(vfslide1up_vf, {})
+CONSTOPT(vfslide1down_vf, {})
+CONSTOPT(vfmerge_vfm, {})
+CONSTOPT(vfmv_v_f, {})
+CONSTOPT(vmfeq_vv, {})
+CONSTOPT(vmfeq_vf, {})
+CONSTOPT(vmfle_vv, {})
+CONSTOPT(vmfle_vf, {})
+CONSTOPT(vmflt_vv, {})
+CONSTOPT(vmflt_vf, {})
+CONSTOPT(vmfne_vv, {})
+CONSTOPT(vmfne_vf, {})
+CONSTOPT(vmfgt_vf, {})
+CONSTOPT(vmfge_vf, {})
+CONSTOPT(vfdiv_vv, {})
+CONSTOPT(vfdiv_vf, {})
+CONSTOPT(vfrdiv_vf, {})
+CONSTOPT(vfmul_vv, {})
+CONSTOPT(vfmul_vf, {})
+CONSTOPT(vfrsub_vf, {})
+CONSTOPT(vfmadd_vv, {})
+CONSTOPT(vfmadd_vf, {})
+CONSTOPT(vfnmadd_vv, {})
+CONSTOPT(vfnmadd_vf, {})
+CONSTOPT(vfmsub_vv, {})
+CONSTOPT(vfmsub_vf, {})
+CONSTOPT(vfnmsub_vv, {})
+CONSTOPT(vfnmsub_vf, {})
+CONSTOPT(vfmacc_vv, {})
+CONSTOPT(vfmacc_vf, {})
+CONSTOPT(vfnmacc_vv, {})
+CONSTOPT(vfnmacc_vf, {})
+CONSTOPT(vfmsac_vv, {})
+CONSTOPT(vfmsac_vf, {})
+CONSTOPT(vfnmsac_vv, {})
+CONSTOPT(vfnmsac_vf, {})
+CONSTOPT(vfwadd_vv, {})
+CONSTOPT(vfwadd_vf, {})
+CONSTOPT(vfwredusum_vs, {})
+CONSTOPT(vfwsub_vv, {})
+CONSTOPT(vfwsub_vf, {})
+CONSTOPT(vfwredosum_vs, {})
+CONSTOPT(vfwadd_wv, {})
+CONSTOPT(vfwadd_wf, {})
+CONSTOPT(vfwsub_wv, {})
+CONSTOPT(vfwsub_wf, {})
+CONSTOPT(vfwmul_vv, {})
+CONSTOPT(vfwmul_vf, {})
+CONSTOPT(vfwmacc_vv, {})
+CONSTOPT(vfwmacc_vf, {})
+CONSTOPT(vfwnmacc_vv, {})
+CONSTOPT(vfwnmacc_vf, {})
+CONSTOPT(vfwmsac_vv, {})
+CONSTOPT(vfwmsac_vf, {})
+CONSTOPT(vfwnmsac_vv, {})
+CONSTOPT(vfwnmsac_vf, {})
